--- docs/modules.md
@@ -1,9 +1,0 @@
-[pluto-encrypted - v0.2.0](README.md) / Exports
-
-# pluto-encrypted - v0.2.0
-
-## Table of contents
-
-### Classes
-
-- [Database](classes/Database.md)
