+++ packages/database/package.json
@@ -21,7 +21,6 @@
     "test:watch": "NODE_ENV=debug vitest tests/*.test.ts",
     "test:debug": "NODE_ENV=debug vitest tests/*.test.ts --inspect-brk --pool threads --poolOptions.threads.singleThread",
     "build": "rm -rf build && npx rollup -c rollup/rollup.mjs",
-    "dev": "rm -rf build && npx rollup -c rollup/rollup.mjs",
     "coverage": "npx vitest run --coverage.enabled --coverage.reporter='json-summary' && npx istanbul-badges-readme --readmeDir=\"../../\"",
     "docs": "npx typedoc --options typedoc.json --hideGenerator"
   },
