+++ packages/database/src/types.ts
@@ -1,5 +1,13 @@
 import { KeyFunctionMap, RxCollectionBase, RxJsonSchema, RxQuery, RxStorageInstanceCreationParams } from "rxdb";
-import { MangoQuery, MangoQueryNoLimit, MangoQuerySelectorAndIndex, RxCollection, RxDocument } from "rxdb/dist/types/types";
+import { MangoQuery, MangoQueryNoLimit, MangoQuerySelectorAndIndex, RxCollection, RxDocument, RxDumpDatabase, RxStorage } from "rxdb/dist/types/types";
+import { CredentialCollection } from "./schemas/Credential";
+import { CredentialRequestMetadataCollection } from "./schemas/CredentialRequestMetadata";
+import { DIDCollection } from "./schemas/DID";
+import { DIDPairCollection } from "./schemas/DIDPair";
+import { LinkSecretColletion } from "./schemas/LinkSecret";
+import { MediatorCollection } from "./schemas/Mediator";
+import { MessageColletion } from "./schemas/Message";
+import { PrivateKeyColletion } from "./schemas/PrivateKey";
 
 
 export type NoKeys<T, K extends keyof T> = {
@@ -18,3 +26,20 @@ export interface GenericORMType<RxDocType> extends RxCollection<any, any, any> {
     remove(queryObj?: MangoQueryNoLimit<any> | string): RxDocument<RxDocType, any>[]
 }
 
+export type PlutoCollections = {
+    messages: MessageColletion;
+    dids: DIDCollection;
+    didpairs: DIDPairCollection;
+    mediators: MediatorCollection;
+    privatekeys: PrivateKeyColletion;
+    credentials: CredentialCollection;
+    credentialrequestmetadatas: CredentialRequestMetadataCollection;
+    linksecrets: LinkSecretColletion;
+};
+
+export type createEncryptedOptions = {
+    name: string,
+    encryptionKey: Uint8Array,
+    importData?: RxDumpDatabase<PlutoCollections>,
+    storage: RxStorage<any, any>
+}
