+++ packages/database/typedoc.json
@@ -14,6 +14,7 @@
     "excludeProtected": true,
     "excludeInternal": true,
     "excludeNotDocumented": true,
+    "readme": "none",
     "excludeNotDocumentedKinds": [
         "TypeAlias",
         "Variable"
